+++ backend/src/main/java/com/eshop/backend/DAO/DataAccess/controllers/AdminController.java
@@ -5,12 +5,12 @@ import com.eshop.backend.dao.Models.AuthorizedUser;
 public interface AdminController extends ManagerController {
     AuthorizedUser getManagerDataByLogin(String login);
     AuthorizedUser updateManagerDataByLogin(String login);
-    AuthorizedUser getManagerDataById(int id);
-    AuthorizedUser updateManagerDataByID(int id);
+    AuthorizedUser getManagerDataById(Long id);
+    AuthorizedUser updateManagerDataByID(Long id);
     AuthorizedUser getCourierDataByLogin(String login);
     AuthorizedUser updateCourierDataByLogin(String login);
-    AuthorizedUser getCourierDataById(int id);
-    AuthorizedUser updateCourierDataById(int id);
+    AuthorizedUser getCourierDataById(Long id);
+    AuthorizedUser updateCourierDataById(Long id);
 
 
 }
