+++ backend/src/main/java/com/eshop/backend/DAO/DataAccess/controllers/ManagerController.java
@@ -6,7 +6,7 @@ import com.eshop.backend.dao.Models.AuthorizedUser;
 public interface ManagerController extends MainDao<AuthorizedUser> {
     AuthorizedUser getManagerDataByLogin(String login);
     AuthorizedUser updateManagerDataByLogin(String login);
-    AuthorizedUser getManagerDataById(int id);
-    AuthorizedUser updateManagerDataByID(int id);
+    AuthorizedUser getManagerDataById(Long id);
+    AuthorizedUser updateManagerDataByID(Long id);
 
 }
